22novPM

+++ ft_putuphex.c
@@ -6,13 +6,13 @@
 /*   By: maroy <maroy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 12:58:34 by maroy             #+#    #+#             */
-/*   Updated: 2022/11/22 13:11:10 by maroy            ###   ########.fr       */
+/*   Updated: 2022/11/22 15:13:12 by maroy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putuphex(int n, int *count)
+void	ft_putuphex(unsigned long n, int *count)
 {
 	if (n > 9)
 	{
